Move analyzer into Analyzers folder
