+++ src/PolyType.SourceGenerator/Analyzers/TypeShapeExtensionAnalyzer.cs
@@ -3,7 +3,7 @@ using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Operations;
 using System.Collections.Immutable;
 
-namespace PolyType.SourceGenerator;
+namespace PolyType.SourceGenerator.Analyzers;
 
 [DiagnosticAnalyzer(LanguageNames.CSharp)]
 public class TypeShapeExtensionAnalyzer : DiagnosticAnalyzer
